v6
Added some error handling if send button is pressed and nothing is being entered to tell to user that Matrikelnummber cannot be empty or contain letters. Then I have added a new button to our UI for next part of the assignment sheet. I have created new class for calculations and created two empty methods for quer summe and int to binary conversation.